Made the Linux delegation strategy OS-agnostic and moved to base module
DelegatingThemeMonitorService has been refactored from LinuxEnvironmentDelegateHelper to be OS-agnostic. Both it and NullThemeMonitorService have been moved to the base module as they are no longer OS-specific.
