+++ linux/src/main/java/com/github/weisj/darkmode/platform/linux/LinuxThemeMonitorService.java
@@ -24,41 +24,22 @@
  */
 package com.github.weisj.darkmode.platform.linux;
 
+import com.github.weisj.darkmode.platform.DelegatingThemeMonitorService;
+import com.github.weisj.darkmode.platform.LibraryUtil;
+import com.github.weisj.darkmode.platform.NullThemeMonitorService;
 import com.github.weisj.darkmode.platform.ThemeMonitorService;
+import com.github.weisj.darkmode.platform.linux.gnome.GnomeThemeMonitorService;
 
-public class LinuxThemeMonitorService implements ThemeMonitorService {
-    @Override
-    public boolean isDarkThemeEnabled() {
-        return LinuxEnvironmentDelegateHelper.isDarkThemeEnabled();
+public class LinuxThemeMonitorService extends DelegatingThemeMonitorService {
+    public LinuxThemeMonitorService() {
+        super(getDelegate());
     }
 
-    @Override
-    public boolean isHighContrastEnabled() {
-        return LinuxEnvironmentDelegateHelper.isHighContrastEnabled();
-    }
-
-    @Override
-    public long createEventHandler(Runnable callback) {
-        return LinuxEnvironmentDelegateHelper.createEventHandler(callback);
-    }
-
-    @Override
-    public void deleteEventHandler(long eventHandle) {
-        LinuxEnvironmentDelegateHelper.deleteEventHandler(eventHandle);
-    }
-
-    @Override
-    public boolean isActive() {
-        return LinuxEnvironmentDelegateHelper.isActive();
-    }
-
-    @Override
-    public void uninstall() {
-        LinuxEnvironmentDelegateHelper.uninstall();
-    }
-
-    @Override
-    public void install() {
-        LinuxEnvironmentDelegateHelper.install();
+    private static ThemeMonitorService getDelegate() {
+        if (LibraryUtil.isGnome) {
+            return new GnomeThemeMonitorService();
+        } else {
+            return new NullThemeMonitorService();
+        }
     }
 }
